update to Swift 5

+++ Sources/BinarySearchTrees/AVLTree.swift
@@ -98,19 +98,19 @@ public enum AVLTree<E: Comparable>: MutableBinarySearchTreeType {
 		let t = tolerance // inbalance tolerance
 		switch self {
 		case let .branch(.branch(.branch(ll, le, lr, llh), e, rl, lh), re, rr, _)
-            where (lh > rr.height + t) && (llh > rl.height + t):
+            where (lh > rr.height + Int8(t)) && (llh > rl.height + Int8(t)):
             // right rotation:
 			return AVLTree(AVLTree(ll, le, lr), e, AVLTree(rl, re, rr))
 		case let .branch(ll, le, .branch(lr, e, .branch(rl, re, rr, rrh), rh), _)
-            where (rh > ll.height + t) && (rrh > lr.height + t):
+            where (rh > ll.height + Int8(t)) && (rrh > lr.height + Int8(t)):
             // left rotation:
 			return AVLTree(AVLTree(ll, le, lr), e, AVLTree(rl, re, rr))
 		case let .branch(.branch(ll, le, .branch(lr, e, rl, lrh), lh), re, rr, _)
-            where (lh > rr.height + t) && (lrh > ll.height + t):
+            where (lh > rr.height + Int8(t)) && (lrh > ll.height + Int8(t)):
             // right left rotation:
 			return AVLTree(AVLTree(ll, le, lr), e, AVLTree(rl, re, rr))
 		case let .branch(ll, le, .branch(.branch(lr, e, rl, rlh), re, rr, rh), _)
-            where (rh > ll.height + t) && (rlh > rr.height + t):
+            where (rh > ll.height + Int8(t)) && (rlh > rr.height + Int8(t)):
             // left right rotation:
 			return AVLTree(AVLTree(ll, le, lr), e, AVLTree(rl, re, rr))
 		default:
